iac/assets/thumbnail_lambda/requirements.txt
boto3-stubs[s3]
pillow
